Renaming response classes, using types for requested configuration.
