+++ graylog2-rest-client/src/main/java/org/graylog2/restclient/models/api/responses/AvailableOutputSummary.java
@@ -18,11 +18,11 @@ package org.graylog2.restclient.models.api.responses;
 
 import com.google.common.collect.Lists;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.graylog2.restclient.lib.plugin.configuration.BooleanField;
-import org.graylog2.restclient.lib.plugin.configuration.DropdownField;
-import org.graylog2.restclient.lib.plugin.configuration.NumberField;
-import org.graylog2.restclient.lib.plugin.configuration.RequestedConfigurationField;
-import org.graylog2.restclient.lib.plugin.configuration.TextField;
+import org.graylog2.rest.models.configuration.responses.BooleanField;
+import org.graylog2.rest.models.configuration.responses.DropdownField;
+import org.graylog2.rest.models.configuration.responses.NumberField;
+import org.graylog2.rest.models.configuration.responses.RequestedConfigurationField;
+import org.graylog2.rest.models.configuration.responses.TextField;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
