+++ graylog2-rest-client/src/test/java/org/graylog2/restclient/models/ConfigurableEntityTest.java
@@ -18,8 +18,8 @@ package org.graylog2.restclient.models;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import org.graylog2.restclient.lib.plugin.configuration.RequestedConfigurationField;
-import org.graylog2.restclient.lib.plugin.configuration.TextField;
+import org.graylog2.rest.models.configuration.responses.RequestedConfigurationField;
+import org.graylog2.rest.models.configuration.responses.TextField;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
