+++ graylog2-rest-models/src/main/java/org/graylog2/rest/models/alarmcallbacks/responses/AvailableAlarmCallbacksResponse.java
@@ -14,10 +14,10 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.restclient.models.api.responses.alarmcallbacks;
+package org.graylog2.rest.models.alarmcallbacks.responses;
 
 import com.google.common.collect.Maps;
-import org.graylog2.restclient.lib.plugin.configuration.RequestedConfigurationField;
+import org.graylog2.rest.models.configuration.responses.RequestedConfigurationField;
 
 import java.util.List;
 import java.util.Map;
@@ -25,13 +25,13 @@ import java.util.Map;
 /**
  * @author Dennis Oelkers <dennis@torch.sh>
  */
-public class GetAvailableAlarmCallbacksResponse {
-    public Map<String, GetSingleAvailableAlarmCallbackResponse> types;
+public class AvailableAlarmCallbacksResponse {
+    public Map<String, AvailableAlarmCallbackSummaryResponse> types;
 
     public Map<String, List<RequestedConfigurationField>> getRequestedConfiguration() {
         Map<String, List<RequestedConfigurationField>> result = Maps.newHashMap();
 
-        for (Map.Entry<String, GetSingleAvailableAlarmCallbackResponse> entry : types.entrySet()) {
+        for (Map.Entry<String, AvailableAlarmCallbackSummaryResponse> entry : types.entrySet()) {
             result.put(entry.getKey(), entry.getValue().extractRequestedConfiguration(entry.getValue().requested_configuration));
         }
