+++ graylog2-rest-models/src/main/java/org/graylog2/rest/models/configuration/responses/BooleanField.java
@@ -14,14 +14,31 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.restclient.models.api.responses.alarmcallbacks;
+package org.graylog2.rest.models.configuration.responses;
 
-import java.util.List;
+import org.graylog2.rest.models.configuration.responses.RequestedConfigurationField;
+
+import java.util.Map;
 
 /**
- * @author Dennis Oelkers <dennis@torch.sh>
+ * @author Lennart Koopmann <lennart@torch.sh>
  */
-public class GetAlarmCallbacksResponse {
-    public Long total;
-    public List<AlarmCallbackSummaryResponse> alarmcallbacks;
+public class BooleanField extends RequestedConfigurationField {
+
+    private final static String TYPE = "boolean";
+
+    public BooleanField(Map.Entry<String, Map<String, Object>> c) {
+        super(TYPE, c);
+    }
+
+    @Override
+    public String getType() {
+        return TYPE;
+    }
+
+    @Override
+    public String attributeToJSValidation(String attribute) {
+        throw new RuntimeException("This type does not have any validatable attributes.");
+    }
+
 }
